apps/tts-svc/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
piper-tts==1.3.0
numpy==1.26.4
python-multipart==0.0.20
requests==2.32.3
